requirements.txt
ddddocr
Pillow